pyproject.toml
[project]
name = "medisolveai-backend"
version = "0.1.0"
description = "MediSolveAI backend assignment multi-module project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.4.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "asyncmy>=0.2.9"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = [
    "Assignment1/tests",
    "Assignment2/tests"
]
